fix(browser): allow importing BrowserCommand if no browser package is installed
