+++ packages/browser/context.d.ts
@@ -168,19 +168,6 @@ export interface ScreenshotMatcherOptions<
   timeout?: number
 }
 
-export interface BrowserCommands {
-  readFile: (
-    path: string,
-    options?: BufferEncoding | FsOptions
-  ) => Promise<string>
-  writeFile: (
-    path: string,
-    content: string,
-    options?: BufferEncoding | (FsOptions & { mode?: number | string })
-  ) => Promise<void>
-  removeFile: (path: string) => Promise<void>
-}
-
 export interface UserEvent {
   /**
    * Creates a new user event instance. This is useful if you need to keep the
@@ -421,7 +408,7 @@ export interface LocatorByRoleOptions extends LocatorOptions {
 
 interface LocatorScreenshotOptions extends Omit<ScreenshotOptions, 'element'> {}
 
-interface LocatorSelectors {
+export interface LocatorSelectors {
   /**
    * Creates a way to locate an element by its [ARIA role](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles), [ARIA attributes](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes) and [accessible name](https://developer.mozilla.org/en-US/docs/Glossary/Accessible_name).
    * @see {@link https://vitest.dev/guide/browser/locators#getbyrole}
