+++ packages/vitest/package.json
@@ -188,7 +188,7 @@
     "tinyglobby": "catalog:",
     "tinypool": "^2.0.0",
     "tinyrainbow": "catalog:",
-    "vite": "^6.0.0 || ^7.0.0-0",
+    "vite": "^6.0.0 || ^7.0.0",
     "why-is-node-running": "^2.3.0"
   },
   "devDependencies": {
@@ -205,7 +205,6 @@
     "@types/picomatch": "^4.0.2",
     "@types/prompts": "^2.4.9",
     "@types/sinonjs__fake-timers": "^8.1.5",
-    "@vitest/browser-playwright": "workspace:*",
     "acorn-walk": "catalog:",
     "birpc": "catalog:",
     "cac": "catalog:",
